fix(v4.1): T2.6 delete dead GoalManager.steerGoal + migrate audit to real steer path
GoalManager.steerGoal had no route and no caller — real goal steering flows
through SessionPrompt.promptOrSteer → steer({delivery:"goal_steer"}). A prior
hardening pass added writeGovernanceAudit + glog to steerGoal, i.e. polished
dead code. Delete steerGoal (interface member, impl, Service.of entry) and move
the governance audit to the live promptOrSteer goal_steer branch where users'
steers actually land.

Extract writeGovernanceAudit into goal-governance-audit.ts (leaf module) so both
goal-manager.editPlan and prompt.ts can use it without a cycle (goal-manager
already imports prompt.ts). New steer.test assertion proves a real goal-steer
writes the audit doc; goal-steer.test's terminal-gate test retargeted to the
shared isTerminalGoalPhase predicate.

3-pkg typecheck 0 error; goal-driver/steer/goal-steer/plan-status-cache green.
